+++ src/Sentry.Unity/SentryUnityOptions.cs
@@ -68,7 +68,11 @@ namespace Sentry.Unity
         /// <summary>
         /// Whether the SDK should add native support for iOS
         /// </summary>
-        public bool IosNativeSupportEnabled { get; set; } = true;
+        public bool IOSNativeSupportEnabled { get; set; } = true;
+        /// <summary>
+        /// Whether the SDK should add native support for Android
+        /// </summary>
+        public bool AndroidNativeSupportEnabled { get; set; } = true;
 
         private readonly SentryMonoBehaviour _sentryMonoBehaviour;