+++ test/Sentry.Unity.Editor.Tests/ScriptableSentryUnityOptionsTests.cs
@@ -37,6 +37,7 @@ namespace Sentry.Unity.Editor.Tests
             StringAssert.Contains("ShutdownTimeout", optionsAsString);
             StringAssert.Contains("MaxQueueItems", optionsAsString);
             StringAssert.Contains("IosNativeSupportEnabled", optionsAsString);
+            StringAssert.Contains("AndroidNativeSupportEnabled", optionsAsString);
             StringAssert.Contains("Debug", optionsAsString);
             StringAssert.Contains("DebugOnlyInEditor", optionsAsString);
             StringAssert.Contains("DiagnosticLevel", optionsAsString);