+++ CHANGELOG.md
@@ -4,12 +4,14 @@
 
 ### Features
 
+- Android Native Support ([#307](https://github.com/getsentry/sentry-unity/pull/307))
 - iOS native bridge for scope sync ([#296](https://github.com/getsentry/sentry-unity/pull/296))
 
 ### Fixes
 
 - Import link.xml caused an infinite loop ([#315](https://github.com/getsentry/sentry-unity/pull/315))
 - Removed unused .asmdefs which clears a warning from console ([#316](https://github.com/getsentry/sentry-unity/pull/316))
+- Don't send negative line number ([#317](https://github.com/getsentry/sentry-unity/pull/317))
 
 ## 0.5.2